product_service/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
gunicorn==22.0.0
